more refactor and placement of classes in their own files
