+++ src/CodeGenerator.cpp
@@ -10,6 +10,7 @@
 #include "ListNode.hpp"
 #include "type.hpp"
 #include "cpptypes.hpp"
+#include "Statement.hpp"
 #include "Compiler.hpp"
 #include "ReturnPacket.hpp"
 #include "symbol_table_stackX.hpp"
