+++ src/cpptypes.cpp
@@ -3,9 +3,11 @@
 #include "type.hpp"
 #include "List.hpp"
 #include "ReturnPacket.hpp"
+#include "TranslationUnit.hpp"
 #include "ParameterBinding.hpp"
 #include "symbol_table_entry.hpp"
 #include "Compiler.hpp"
+
 namespace njnr
 {
    ReturnPacket::ReturnPacket() : m_pair{},
