+++ src/include/common/symbol_table_stack.hpp
@@ -6,10 +6,11 @@
 #include <string>
 
 #include "symbol_table.hpp"       // for class Table
-#include "Compiler.hpp"           // for class Compiler
 
 namespace njnr
 {
+class Compiler;
+
 class SymbolTable
 {
    public:
