+++ src/statements/statements.cpp
@@ -9,6 +9,7 @@
 #include "List.hpp"
 #include "ListNode.hpp"
 #include "ReturnPacket.hpp"
+#include "Statement.hpp"
 
 namespace njnr
 {
